Data-Pipeline/requirements.txt
numpy==2.2.2
pandas==2.2.3
yfinance
matplotlib==3.10.0
google-cloud-storage==3.0.0
google-api-core==2.24.1
tqdm==4.67.1
polars==1.22.0
python-dotenv==1.0.1
fastexcel==0.12.1
coverage==7.6.12
openpyxl==3.1.5
RapidFuzz==3.12.1
scipy==1.15.2
apache-airflow
XlsxWriter==3.2.2